refactor: move back the main processing loop to lib.rs
